Added metadata synching

+++ migrateRepo.sh
@@ -2,6 +2,8 @@
 
 SCRIPT_HOME="$(realpath $(dirname $0))"
 BFG_JAR="$SCRIPT_HOME/bfg.jar"
+METADATA_SYNC_HOME="$SCRIPT_HOME/node-gitlab-2-github"
+METADATA_SYNC_SETTING_TEMPLATE="$SCRIPT_HOME/meta_sync_setting_template.ts"
 configFile=$HOME/.ssh/MigrationConfig
 size_cutoff=49 # max size in MB before we need to lfs.
 
@@ -12,7 +14,7 @@ if [ "$#" -ne 1 ];then
 	echo "       is found in '$configFile' which must have permissions of '600'."
 	exit
 else
-	repo="$1"
+	REPO="$1"
 fi
 if [ ! -f $configFile ];then
 	echo "Aborting: Cannot find config file '$configFile'"
@@ -28,7 +30,7 @@ else
 	. $configFile
 fi
 
-if [ hash gh >& /dev/null ];then
+if [ $(hash gh >& /dev/null;echo $?) -ne 0 ];then
 	echo "Aborting:  The Git Hub CLI 'gh' must be installed in the scripts path."
 	echo "For installation instructions see: "
 	echo "https://computingforgeeks.com/how-to-install-github-cli-on-linux-and-windows"
@@ -47,7 +49,20 @@ if [ $(java -h >& /dev/null;echo $?) -ne 0 ];then
 	echo "https://java.com/en/download/help/download_options.html"
 	exit -1
 fi
+if [ $(hash jq >& /dev/null;echo $?) -ne 0 ];then
+	echo "Aborting:  The command-line JSON processor 'jq' must be installed in"
+	echo "this scripts path. For installation instructions see: "
+	echo "https://stedolan.github.io/jq/"
+	exit -1
+fi
 
+
+if [ ! -f $BFG_JAR ];then
+	echo "Aborting:  BFG Repo-Cleaner jar file $BFG_JAR not found."
+	echo "Use wget to get latest version from: "
+	echo "https://repo1.maven.org/maven2/com/madgag/bfg/"
+	exit -1
+fi
 
 if [ ! -f $BFG_JAR ];then
 	echo "Aborting:  BFG Repo-Cleaner jar file $BFG_JAR not found."
@@ -58,21 +73,22 @@ fi
 
 
 
-echo "Migration: $GL_HOST:$GL_ORG/$repo => $GH_HOST/$GH_ORG"
+echo "Migration: $GL_HOST:$GL_ORG/${REPO} => $GH_HOST/$GH_ORG"
 wrkingDir=$(realpath $PWD)
 time_stamp=$(date +'%Y%m%d%H%m%s')
 
-gl_connect_log="${wrkingDir}/${repo}.gl_connect.${time_stamp}.log"
+gl_connect_log="${wrkingDir}/${REPO}.gl_connect.${time_stamp}.log"
+GL_REPO_ID=$(curl -s -k --header "Authorization: Bearer xCVsx4vgD37CbZCZGRQb" "https://nwcgit.nwfsc.noaa.gov/api/v4/groups/fram-data/search?scope=projects&search=${REPO}"|jq '.[]| select( .path == "'${REPO}'").id') 2>>$gl_connect_log
 
-git ls-remote ${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/$repo >> $gl_connect_log 2>&1
-if [ $? -ne 0 ];then
-	echo "      Aborting: Unable to access '${GL_HOST}:${GL_ORG}/$repo' as '${GL_GIT_USER}'"
+#git ls-remote ${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/${REPO} >> $gl_connect_log 2>&1
+if [ "$GL_REPO_ID" = "" ];then
+	echo "      Aborting: Unable to access '${GL_HOST}:${GL_ORG}/${REPO}'"
 	if [ -r $gl_connect_log ];then
 		cat $gl_connect_log |fold -w 80| sed 's+^+      +'
 	fi
 	exit -1
 fi
-echo "   - Confirmed git access '${GL_HOST}:${GL_ORG}/$repo'"
+echo "   - Confirmed git access '${GL_HOST}:${GL_ORG}/${REPO}'"
 rm -f $gl_connect_log
 
 #Verify GitHub Access login
@@ -84,38 +100,38 @@ fi
 echo "   - Confirmed Git and GitHub CLI (gh) access to '${GH_HOST}'"
 
 #Create or overright remote repo
-target_repo_log="${wrkingDir}/${repo}.target_repo.${time_stamp}.log"
+target_repo_log="${wrkingDir}/${REPO}.target_repo.${time_stamp}.log"
 
-gh repo view ${GH_ORG}/$repo >> $target_repo_log 2>&1
+gh repo view ${GH_ORG}/${REPO} >> $target_repo_log 2>&1
 if [ $? -eq 0 ];then
-	echo "   - Deleting existing repo ${GH_GIT_USER}@${GH_HOST}:${GH_ORG}/$repo"
+	echo "   - Deleting existing repo ${GH_GIT_USER}@${GH_HOST}:${GH_ORG}/${REPO}"
 	out=$(mktemp)
 	curl -s -X DELETE \
 		 -H "Accept: application/vnd.github.v3+json" \
 		 -H "Authorization: token ${GH_TOKEN}" \
-		 https://${GH_HOST}/api/v3/repos/${GH_ORG}/$repo --output $out  >> $target_repo_log 2>&1
+		 https://${GH_HOST}/api/v3/repos/${GH_ORG}/${REPO} --output $out  >> $target_repo_log 2>&1
 	if [ $(cat $out|wc -l) -ne 0 ];then
-		echo "      Aborting: Failed to delete ${GH_GIT_USER}@${GH_HOST}:${GH_ORG}/$repo"
+		echo "      Aborting: Failed to delete ${GH_GIT_USER}@${GH_HOST}:${GH_ORG}/${REPO}"
 		if [ -r $out ];then
 			cat $out |fold -w 80 | sed 's+^+      +'
 		fi
 		exit -1
 	fi
-	rm $out
+	rm -f $out
 
 fi
 
-echo "   - Creating repository '${GH_HOST}:${GH_ORG}/$repo'"
+echo "   - Creating repository '${GH_HOST}:${GH_ORG}/${REPO}'"
 # Clean up for cloning
-rm -r -f $repo
+rm -r -f ${REPO}
 
 #NB Sometimes hg repo create returns bac value even after successful creation
 #So check with gh repo view
 
-gh repo create -y  -t ${GH_TEAM} --public ${GH_ORG}/$repo >> $target_repo_log 2>&1
-gh repo view ${GH_ORG}/$repo  >> $target_repo_log 2>&1
+gh repo create -y  -t ${GH_TEAM} --public ${GH_ORG}/${REPO} >> $target_repo_log 2>&1
+gh repo view ${GH_ORG}/${REPO}  >> $target_repo_log 2>&1
 if [ $? -ne 0 ];then
-	echo "      Aborting '$repo' migration: Could not create '${GH_HOST}:${GH_ORG}/$repo'"
+	echo "      Aborting '${REPO}' migration: Could not create '${GH_HOST}:${GH_ORG}/${REPO}'"
 	if [ -r $target_repo_log ];then
 		cat $target_repo_log |fold -w 80| sed 's+^+      +'
 	fi
@@ -124,17 +140,17 @@ fi
 rm -f $target_repo_log
 
 
-echo "   - Creating local clone of '${GL_HOST}:${GL_ORG}/$repo'"
-clone_log="${wrkingDir}/${repo}.clone.${time_stamp}.log"
+echo "   - Creating local clone of '${GL_HOST}:${GL_ORG}/${REPO}'"
+clone_log="${wrkingDir}/${REPO}.clone.${time_stamp}.log"
 
-rm -f -r $repo  # cleanup
+rm -f -r ${REPO}  # cleanup
 
-mkdir $repo
-cd $repo
+mkdir ${REPO}
+cd ${REPO}
 
-git clone ${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/$repo .	>>$clone_log 2>&1
+git clone ${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/${REPO} .	>>$clone_log 2>&1
 if [ $? -ne 0 ];then
-	echo "      Aborting local clone of '${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/$repo'"
+	echo "      Aborting local clone of '${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/${REPO}'"
 	if [ -r $clone_log ];then
 		cat $clone_log |fold -w 80 | sed 's+^+      +'
 	fi
@@ -147,7 +163,7 @@ fi
 	  git fetch --all --prune --tags && git pull --all) 	>>$clone_log 2>&1
 
 if [ $? -ne 0 ];then
-	echo "      Aborting: Could not recover all branches of '${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/$repo'"
+	echo "      Aborting: Could not recover all branches of '${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/${REPO}'"
 	if [ -r $clone_log ];then
 		cat $clone_log |fold -w 80 | sed 's+^+      +'
 	fi
@@ -156,13 +172,10 @@ fi
 rm -f  $clone_log
 
 
-# Protect origin from accidents, and create a purely local repo, which helps lfs
-git remote remove origin 
-
-lfs_log="${wrkingDir}/${repo}.lsf.${time_stamp}.log"
+lfs_log="${wrkingDir}/${REPO}.lsf.${time_stamp}.log"
 git lfs install --local >> $lfs_log 2>&1
 
-tracked="$(git lfs migrate info --above=${size_cutoff}Mb 2> /dev/null |sed 's+\s.*++'|tr '\n' ' ' )"
+tracked="$(git lfs migrate info --everything --above=${size_cutoff}Mb 2> /dev/null |sed 's+\s.*++'|tr '\n' ' ' )"
 
 if [ "$tracked" = "" ];then
 	echo "   - LFS preprocessing skipped: no large (>${size_cutoff}MB) files found."
@@ -173,8 +186,8 @@ else
 		git checkout $branch >>$lfs_log 2>&1
 		if [ $? -ne 0 ];then
 			echo ""
-			echo "       Aborting LFS migration on $repo' branch '$branch': "
-			echo "       LFS migration: unable to check out branch."
+			echo "       Aborting LFS migration on ${REPO}' branch '$branch': "
+			echo "       LFS migration: unable to check out branch '$branch'."
 			if [ -r $lfs_log ];then
 				cat $lfs_log |fold -w 80| sed 's+^+      +'
 			fi
@@ -187,7 +200,7 @@ else
 			 git commit -am "Adding .gitattributes" ) >> $lfs_log 2>&1
 		if [ $? -ne 0 ];then
 				echo ""
-				echo "       Aborting LFS migration on $repo' branch '$branch': "
+				echo "       Aborting LFS migration on ${REPO}' branch '$branch': "
 				echo "       Unable to set up lfs tracking'."
 				if [ -r $lfs_log ];then
 					cat $lfs_log |fold -w 80| sed 's+^+      +'
@@ -202,7 +215,7 @@ else
 			 --convert-to-git-lfs $bfg_list --no-blob-protection $PWD >>$lfs_log 2>&1
 		if [ $? -ne 0 ];then
 			echo ""
-			echo "       Aborting LFS migration on $repo' branch '$branch': "
+			echo "       Aborting LFS migration on ${REPO}' branch '$branch': "
 			echo "       java -jar ${BFG_JAR} failed"
 			if [ -r $lfs_log ];then
 				cat $lfs_log |fold -w 80| sed 's+^+      +'
@@ -213,7 +226,7 @@ else
 	(git reflog expire --expire=now --all  && \
 		 git gc --prune=now --aggressive)  >> $lfs_log 2>&1
 		if [ $? -ne 0 ];then
-			echo "       Aborting LFS migration on $repo' branch '$branch': "
+			echo "       Aborting LFS migration on ${REPO}' branch '$branch': "
 			echog"       git clean/garbage collection failed."
 			if [ -r $lfs_log ];then
 				cat $lfs_log |fold -w 80| sed 's+^+      +'
@@ -224,12 +237,11 @@ else
 fi
 rm -f $lfs_log
 
-echo "   - Pushing to ${GH_HOST}:${GH_ORG}/$repo"
-push_log="${wrkingDir}/${repo}.push.${time_stamp}.log"
-git remote add gitlab "${GL_GIT_USER}@${GL_HOST}:${GL_ORG}/$repo"
-git remote add github "${GH_GIT_USER}@${GH_HOST}:${GH_ORG}/$repo" 
+echo "   - Pushing to ${GH_HOST}:${GH_ORG}/${REPO}"
+push_log="${wrkingDir}/${REPO}.push.${time_stamp}.log"
+git remote add github "${GH_GIT_USER}@${GH_HOST}:${GH_ORG}/${REPO}" 
 
-git config lfs.https://${GH_HOST}/${GH_ORG}/${repo}.git/info/lfs.locksverify true
+git config lfs.https://${GH_HOST}/${GH_ORG}/${REPO}.git/info/lfs.locksverify true
 
 git lfs push --all github >> $push_log 2>&1
 if [ $? -ne 0 ];then
@@ -260,6 +272,31 @@ if [ $? -ne 0 ];then
 	fi
 	exit -1
 fi
+echo "   - Synching metadata (issues, pull requests, etc)"
+meta_log="${wrkingDir}/${REPO}.meta.${time_stamp}.log"
+cd "$METADATA_SYNC_HOME"
+npm install >> $meta_log 2>&1
+cat "$METADATA_SYNC_SETTING_TEMPLATE" \
+	| sed 's+{{GL_HOST}}+'$GL_HOST'+'\
+	| sed 's+{{GH_HOST}}+'$GH_HOST'+'\
+	| sed 's+{{GL_TOKEN}}+'$GL_TOKEN'+'\
+	| sed 's+{{GH_TOKEN}}+'$GH_TOKEN'+'\
+	| sed 's+{{GL_REPO_ID}}+'$GL_REPO_ID'+'\
+	| sed 's+{{GL_ORG}}+'$GL_ORG'+'\
+	| sed 's+{{GH_ORG}}+'$GH_ORG'+'\
+	| sed 's+{{REPO}}+'$REPO'+' >settings.ts
+
+npm run start >> $meta_log 2>&1
+
+if [ $? -ne 0 ];then
+	echo "      Aborting Metadata copy:"
+	if [ -r $meta_log ];then
+		cat $meta_log |fold -w 80 | sed 's+^+      +'
+	fi
+	exit -1
+fi
+rm -f settings.ts
+rm -f $meta_log
 rm -f $push_log
 cd $wrkingDir
 rm -f $gl_connect_log
@@ -267,8 +304,8 @@ rm -f $target_repo_log
 rm -f $clone_log
 rm -f $lfs_log
 rm -f $push_log
-rm -f -r $repo
-rm -f -r $repo.bfg-report
+rm -f -r ${REPO}
+rm -f -r ${REPO}.bfg-report
 echo "Migration completed."
 exit 0
